DRTII-1567 Start refactoring to fit min-staff update service
